SysLogBLL 改为model

+++ mvc.BLL/SysLogBLL.cs
@@ -3,6 +3,8 @@ using mvc.Common;
 using mvc.IBLL;
 using mvc.IDAL;
 using mvc.Models;
+using mvc.Models.Sys;
+using mvc.BLL.Core;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,12 +17,10 @@ namespace mvc.BLL
     {
         [Dependency]
         public ISysLogRepository logRepository { get; set; }
+        DBContainer db = new DBContainer();
 
-
-        public List<SysLog> GetList(ref GridPager pager, string queryStr)
+        public List<SysLogModel> GetList(ref GridPager pager, string queryStr)
         {
-            DBContainer db = new DBContainer();
-            List<SysLog> query = null;
             IQueryable<SysLog> list = logRepository.GetList(db);
             if (!string.IsNullOrWhiteSpace(queryStr))
             {
@@ -34,37 +34,84 @@ namespace mvc.BLL
 
             if (pager.order == "desc")
             {
-                query = list.OrderByDescending(c => c.CreateTime).Skip((pager.page - 1) * pager.rows).Take(pager.rows).ToList();
+                list = list.OrderByDescending(c => c.CreateTime);
             }
             else
             {
-                query = list.OrderBy(c => c.CreateTime).Skip((pager.page - 1) * pager.rows).Take(pager.rows).ToList();
+                list = list.OrderBy(c => c.CreateTime);
             }
 
 
-            return query;
-        }
-        public SysLog GetById(string id)
-        {
-            return logRepository.GetById(id);
+            return CreateModelList(ref pager, ref list);
         }
 
-        public bool Delete(string id)
+        public List<SysLogModel> CreateModelList(ref GridPager pager, ref IQueryable<SysLog> list)
+        {
+            if (pager.page <= 1)
+            {
+                list = list.Take(pager.rows);
+            }
+            else
+            {
+                list = list.Skip((pager.page - 1) * pager.rows).Take(pager.rows);
+            }
+            List<SysLogModel> modelList = (from r in list
+                                           select new SysLogModel
+                                           {
+                                               Id = r.Id,
+                                               Operator = r.Operator,
+                                               Message = r.Message,
+                                               Result = r.Result,
+                                               Type = r.Type,
+                                               Module = r.Module,
+                                               CreateTime = r.CreateTime
+
+                                           }).ToList();
+            return modelList;
+        }
+
+        public SysLogModel GetById(string id)
+        {
+            SysLog entity = logRepository.GetById(id);
+            SysLogModel model = new SysLogModel();
+            if (entity == null)
+                return model;
+            model.Id = entity.Id;
+            model.Message = entity.Message;
+            model.Module = entity.Module;
+            model.Operator = entity.Operator;
+            model.Result = entity.Result;
+            model.Type = entity.Type;
+            model.CreateTime = entity.CreateTime;
+            return model;
+        }
+
+        public bool Delete(ref ValidationErrors errors, string id)
         {
             try
             {
-                DBContainer db = new DBContainer();
-                if (logRepository.Delete(db,id) == 1)
+                if (!string.IsNullOrWhiteSpace(id))
                 {
-                    return true;
+                    if (logRepository.Delete(db, id) == 1)
+                    {
+                        return true;
+                    }
+                    else
+                    {
+                        errors.Add("删除失败");
+                        return false;
+                    }
                 }
                 else
                 {
+                    errors.Add("主键重复");
                     return false;
                 }
             }
             catch (Exception ex)
             {
+                errors.Add(ex.Message);
+                ExceptionHander.WriteException(ex);
                 return false;
             }
         }
